fix: fix logic and deadlock in room
